Fix self-imports to use correct package root
While imports from CWD work, the "ipsiblings" directory
is marked as a Python package using the __init__.py
file. Therefore, the root package of the application
is ipsiblings and not no package. Hence, the correct
way to import files from our own package is to import
them from . or .. .

The previous state (correctly) caused compiler errors
in IntelliJ.

https://docs.python.org/3/tutorial/modules.html#intra-package-references