Remove manual patching of scipy libarary
No longer necessary as of scipy 1.4.0, the bug mentioned
in settings.py has since been resolved. The library
version is now fixed using requirements.txt, so this
should no longer be an issue.

Also, the source code has changed, mandating an
update of the patch condition. This is technically
no longer necessary and might be wise to remove
though.